A LOT of miscelaneous changes!
Including but not limited to, permanently enabling some bugfixes, re-labelling some things according to Sonic 1/2 standards, renaming a constant, and replacing the nemesis compressor with a better one. Same for Object 09 -- not only labelled, the jump bugfix was ported. Special stages still unplayable tho. And a temp folder too, for assets to be used soon:tm: