Game over hotfix
Until we have the continue screen re-implemented, we'll handle continues here. Since we can't send the player to GamemodeID_Level (it'd just go back to the title screen, in this case), we consume the continue, send them to a special stage, and set their lives counter to 1.

A workaround; but this could actually be turned into a feature; a "LAST CHANCE!" Special stage with different coloration; find the 1UP token, and you're back! Fail...

And it's GAME OVER. This time, for real.