Undo the change of nemesis to Twizzler
turns out the buffer is bigger than anticipated. $2000 bytes. I can not afford that; it gets in the way of future plans. I'm reverting temporarily to Nemesis -- KosM will be the way to go. (hence why some Nemesis files in the kosinski folder -- logic in the chaos.)

Twizzler will not be Ditched however: Regular Twizzler Decompression is still useful. Moduled Twizzler will be trimmed out, tho: That's essentially dead code.