Cleanup
Deleted some leftovers from the previous driver & the porting process. I've also re-introduced some macros that were erroneously split before. And re-arranged the music to follow Sonic 2's level order (Not that it matters since that's not the FINAL level order, but hey, it's organized -- a good starting point to re-shuffle it once this level's romhack get properly introduced.