Misc. adjustments
I've decided that the player will NOT bsr to objectMove/objectMoveAndFall -- it will have this code duplicated inside of it. Maximum speed at the cost of negligible space cost. The GHZ Edge walls have been fixed to behave as intended -- It now uses MarkObjGone instead of the macro. (Speaking of, Obj16 now follows the same behaviour; should be faster than just having a copy with two jumps, to display sprite and Delete object).