Musical update
Not much to say, really :P
Aside from the fact that the boss music is now that of Sonic 2's beta. Sort of. I am experimenting with the values, for fun:tm:
Also, duplicated some code in instances where a "bra.w" was used, in the collision subroutines: Collision is important, so may as well make a micro-optimization ;P
Also optimized the Seesaw object: After juggling the code around back and forth (it's now what, the 3rd time?), it's now a beautiful, concise block~ Or at least, part of it is.