The calm before the storm
Changelog? The answer is yes. Too many to count. But: it's mostly optimizations, correcting comments, and reorganizing a few assets. (Such the re-ordering of some objects, which begins in this commit). A few bugfixes too from historical guides (ground/air speecap), along with the start of the Homming attack code -- I will get to completing this once I migrate statuses (shield, inv, etc.) to bitfields, alike sonic 2 final.

Also (to be finished): Turning obTimeFrame from word to byte, to free an universal SST.