Simplify a few calculations
And whilst we're at it -- Sonic 1's ring art is king again~ The older file is no longer needed. With that, rings look proper again.
As an aside, we're also fixing object positioning for HPZ_1 -- it broke when Github desktop had an aneurism. Recovering it was just as easy, tho.